MSPSDS-209 Move duplicate searching code into a concern
